requirements_sd3.txt
accelerate>=0.31.0
torchvision
transformers>=4.41.2
ftfy
tensorboard
Jinja2
peft== 0.11.1
wandb
